Expand SpellEffects in enums.py
Added support for 2 spell effects, 147 (extends a damage over time), and 148 (extends a heal over time). 
You can find examples of SpellEffect 147 in the spell "Blast Off!" or 147 and 148 (presumably) in the item card spell 'Duck Savage".
NOTE: 148 is inferred, not confirmed. 147 is confirmed to be ending a damage over time though.